requirements.txt
absl-py==0.9.0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.1.2
asn1crypto==0.24.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.2
Automat==0.7.0
Babel==2.7.0
backcall==0.2.0
base32hex==1.0.2
bcrypt==3.1.7
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.0
certifi==2019.6.16
cffi==1.12.3
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
choice==0.1
chronometry==2020.4.28.2
click==7.1.2
colorama==0.4.3
colouration==2020.4.28.2
constantly==15.1.0
cryptography==2.8
cv2wrap==1.0
cycler==0.10.0
daphne==2.4.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
disk==2020.4.28.2
distlib==0.3.0
distro==1.5.0
dj-config-url==0.1.1
Django==2.2.3
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.6.0
django-colorfield==0.1.15
django-cookies-samesite==0.5.1
django-cors-headers==3.2.1
django-countries==6.1.2
django-filter==2.2.0
django-hashid-field==1.0.0
django-mathfilters==1.0.0
django-model-utils==4.0.0
django-oauth-toolkit==1.2.0
django-otp==0.7.5
django-paypal==1.0.0
django-phonenumber-field==4.0.0
django-picklefield==3.0.1
django-qr-code==1.1.0
django-social-authsch==0.1
django-twilio==0.11.0
django-user-agents==0.4.0
django-wkhtmltopdf==3.3.0
djangorestframework==3.11.0
easy-thumbnails==2.7
filelock==3.0.12
func-timeout==4.3.5
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
geographiclib==1.50
geopy==1.22.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hashids==1.2.0
hiredis==1.0.0
html2pdf==0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.8.0
importlib-metadata==1.6.0
importlib-resources==1.5.0
imutils==0.5.3
incremental==17.5.0
ipython==7.15.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.17.1
joblib==0.15.1
Keras==2.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.2
matplotlib==3.2.2
msgpack==0.6.2
msgpack-python==0.5.6
mysql==0.0.2
mysqlclient==1.4.6
networkx==2.4
now==0.0.6
numpy==1.18.1
oauth-provider==1.1
oauth2-provider==0.0
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.1
packaging==20.3
paramiko==2.7.1
parso==0.7.0
path==14.0.1
paypal==1.2.5
pdf==2020.4.28.2
pdfkit==0.6.1
pdfminer.six==20200517
pdftotree==0.4.0
pexpect==4.8.0
phonenumbers==8.11.2
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
progressbar2==3.51.3
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2-binary==2.8.6
ptyprocess==0.6.0
public==2019.4.13
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.7
Pygments==2.6.1
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-decouple==3.1
python-memcached==1.59
python-social-auth==0.3.6
python-utils==2.4.0
python3-openid==3.1.0
pytz==2019.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzbar==0.1.8
qrcode==6.1
ravenclaw==2020.4.28.2
rdflib==5.0.0
redis==2.10.6
regex==2020.6.8
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
rm==2019.4.13
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
selectivesearch==0.4
Send2Trash==1.5.0
service-identity==18.1.0
silverware==2020.4.28.2
six==1.12.0
sklearn==0.0
slytherin==2020.4.28.2
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SPARQLWrapper==1.8.5
sqlparse==0.3.0
standard==1.0.3
stripe==2.42.0
swapper==1.1.2.post1
tabula-py==2.1.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.0
tox==3.15.0
tqdm==4.46.1
traitlets==4.3.3
twilio==6.35.2
Twisted==19.10.0
txaio==18.8.1
ua-parser==0.8.0
Unidecode==1.1.1
urllib3==1.25.7
user-agents==2.1
values==2019.4.13
virtualenv==20.0.20
Wand==0.6.1
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
xhtml2pdf==0.2.3
zipp==3.1.0
zope.interface==5.1.0
